Minor fixes to enemy list and game start

+++ AssassinValkyrie/cursor.h
@@ -1,6 +1,6 @@
 // Module			: Gameplay Programming
 // Assignment2		: Assassin Valkyrie
-// Student Number	: Chua Wei trooper Timothy
+// Student Number	: Chua Wei Jie Timothy
 // Student Number	: S10165581F
 
 #ifndef _CURSOR_COMPONENT_H
@@ -15,8 +15,8 @@ namespace cursorNS
 	const int	HEIGHT = 32;
 	const double ROTATION = 0;
 	const float SCALE = 1.0f;
-	const int	X = GAME_WIDTH / 2 - WIDTH*SCALE;
-	const int	Y = GAME_HEIGHT / 2 - HEIGHT*SCALE;
+	const float	X = GAME_WIDTH / 2 - WIDTH*SCALE;
+	const float	Y = GAME_HEIGHT / 2 - HEIGHT*SCALE;
 	const int   TEXTURE_COLS = 1;
 	const int   START_FRAME = 0;
 	const int   END_FRAME = 0;
